Initial introduction of jetty-core-webapp
