+++ jetty-core/jetty-bom/pom.xml
@@ -52,6 +52,11 @@
       </dependency>
       <dependency>
         <groupId>org.eclipse.jetty</groupId>
+        <artifactId>jetty-core-webapp</artifactId>
+        <version>12.1.0-SNAPSHOT</version>
+      </dependency>
+      <dependency>
+        <groupId>org.eclipse.jetty</groupId>
         <artifactId>jetty-deploy</artifactId>
         <version>12.1.0-SNAPSHOT</version>
       </dependency>
