+++ jetty-core/jetty-core-test-webapps/jetty-core-bad-init-webapp/src/main/assembly/core-webapp.xml
@@ -1,0 +1,39 @@
+<assembly xmlns="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.0 http://maven.apache.org/xsd/assembly-1.1.0.xsd">
+  <id>core-webapp</id>
+  <formats>
+    <format>zip</format>
+  </formats>
+  <includeBaseDirectory>false</includeBaseDirectory>
+  <fileSets>
+    <fileSet>
+      <directory>${basedir}/${project.build.outputDirectory}</directory>
+      <outputDirectory>/classes</outputDirectory>
+      <includes>
+        <include>**/*.*</include>
+      </includes>
+      <excludes>
+        <exclude>jetty-core.xml</exclude>
+      </excludes>
+    </fileSet>
+    <fileSet>
+      <directory>${basedir}/${project.build.outputDirectory}</directory>
+      <outputDirectory>/</outputDirectory>
+      <includes>
+        <include>jetty-core.xml</include>
+      </includes>
+    </fileSet>
+    <fileSet>
+      <directory>${basedir}/src/main/static</directory>
+      <outputDirectory>/static</outputDirectory>
+      <includes>
+        <include>**/*.*</include>
+      </includes>
+    </fileSet>
+  </fileSets>
+  <dependencySets>
+    <dependencySet>
+      <outputDirectory>/lib</outputDirectory>
+      <useProjectArtifact>false</useProjectArtifact>
+    </dependencySet>
+  </dependencySets>
+</assembly>
