+++ jetty-core/jetty-core-test-webapps/jetty-core-bad-init-webapp/src/main/java/org/example/ExampleBadStartHandler.java
