+++ jetty-core/jetty-core-test-webapps/pom.xml
@@ -1,25 +1,26 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <groupId>org.eclipse.jetty</groupId>
-    <artifactId>jetty-core</artifactId>
+    <artifactId>jetty-project</artifactId>
     <version>12.1.0-SNAPSHOT</version>
+    <relativePath>../../pom.xml</relativePath>
   </parent>
-  <artifactId>jetty-tests</artifactId>
+
+  <groupId>org.eclipse.jetty.tests</groupId>
+  <artifactId>jetty-core-test-webapps</artifactId>
   <packaging>pom</packaging>
-  <name>Core :: Tests</name>
+
+  <name>Core :: WebApp :: Test WebApps</name>
+  <description>Jetty Core WebApp :: Test WebApps</description>
 
   <modules>
-    <module>jetty-test-rfc</module>
-    <module>jetty-test-client-transports</module>
-    <module>jetty-test-jmx</module>
-    <module>jetty-test-http2-client-transport-provided-webapp</module>
+    <module>jetty-core-bad-init-webapp</module>
+    <module>jetty-core-demo-webapp</module>
   </modules>
 
   <properties>
     <maven.deploy.skip>true</maven.deploy.skip>
   </properties>
-
 </project>
