+++ jetty-core/jetty-core-webapp/src/test/resources/jetty-logging.properties
@@ -1,0 +1,9 @@
+# Jetty Logging using jetty-slf4j-impl
+#org.eclipse.jetty.deploy.DeploymentTempDirTest.LEVEL=DEBUG
+#org.eclipse.jetty.deploy.LEVEL=DEBUG
+#org.eclipse.jetty.ee9.webapp.LEVEL=DEBUG
+#org.eclipse.jetty.util.Scanner=DEBUG
+org.eclipse.jetty.server.Server.LEVEL=WARN
+org.eclipse.jetty.server.AbstractConnector.LEVEL=WARN
+org.eclipse.jetty.server.handler.ContextHandler.LEVEL=WARN
+org.eclipse.jetty.util.ssl.SslContextFactory.LEVEL=WARN
