--- jetty-core/jetty-server/src/main/config/etc/jetty-core-deploy.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0"?><!DOCTYPE Configure PUBLIC "-//Jetty//Configure//EN" "https://jetty.org/configure_10_0.dtd">
-
-<!-- =============================================================== -->
-<!-- Configure the "core" environment deployment defaults            -->
-<!-- =============================================================== -->
-<Configure>
-  <Ref refid="deploymentScanner">
-    <Call name="configureEnvironment">
-      <Arg>core</Arg>
-      <Set name="classLoaderFactoryClassName">
-        <Property name="jetty.deploy.classLoaderFactoryClass" default="org.eclipse.jetty.server.handler.CoreContextHandler$CoreContextClassLoaderFactory" />
-      </Set>
-      <Set name="defaultContextHandlerClassName">
-        <Property name="jetty.deploy.defaultContextHandlerClass" default="org.eclipse.jetty.server.handler.CoreContextHandler" />
-      </Set>
-    </Call>
-  </Ref>
-</Configure>
