--- jetty-core/jetty-server/src/main/config/modules/core-deploy.mod
@@ -1,24 +1,0 @@
-[description]
-Scans and deploys Jetty `core` environment web applications.
-
-[tags]
-deployment
-
-[environment]
-core
-
-[before]
-static-deploy
-
-[depend]
-deployment-scanner
-
-[xml]
-etc/jetty-core-deploy.xml
-
-[ini-template]
-## ClassLoaderFactory class for core environment deployments
-# jetty.deploy.classLoaderFactoryClass=org.eclipse.jetty.server.handler.CoreContextHandler$CoreContextClassLoaderFactory
-
-## Default ContextHandler class for core environment deployments
-# jetty.deploy.defaultContextHandlerClass=org.eclipse.jetty.server.handler.CoreContextHandler
