--- jetty-core/jetty-server/src/main/config/modules/static-deploy.mod
@@ -1,18 +1,0 @@
-[description]
-Scans and deploys Jetty `static` environment web applications.
-
-[tags]
-deployment
-
-[environment]
-static
-
-[depend]
-deployment-scanner
-
-[xml]
-etc/jetty-static-deploy.xml
-
-[ini-template]
-## Default ContextHandler class for static environment deployments
-# jetty.deploy.defaultContextHandlerClass=org.eclipse.jetty.server.handler.StaticContextHandler
